+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 * Added Waveform widget move both dual markers when click on top Δt bar  #247
 * Added preferred units to Waveform widget i, v, p plots  #248
 * Added time format options to CSV statistics recording  #231
+* Added option to set exact plot y-axis range in Waveform widget  #126
 
 
 ## 1.0.48